[dev] `/healthz`
We currently have some races that cause some of the test cases to fail
because the server takes several seconds to warm up and be truly ready.

This change adds a `/healthz` endpoint that responds to HTTP so that the
Docker driver can accurately wait until the container is ready. It also
stops making the Docker containers autoremovable, because that causes
some systems to tear down the container too fast and they cannot be
restarted effectively.