Better distinguish between server and client actions
The action types are currently a bit of a mess because we rely on
`Exclude` to build the concrete type. But why do all that gymnastics
when we can simply declare the type as expected?

This change stops the madness and declares the `ServerAction` and
`ClientAction` types separately, and renames `wait` to `sleep` for
clarity.