Wait for messages, not time at the end
Right now the test harness always waits a constant 2s after all messages
have been sent. We don't want to be in a situation where we are waiting
for constant time and instead should be waiting for things to be
observed.

This change makes it such that we now wait for all the expected outputs
to arrive, instead of waiting a constant 2s.