Add `wait_response` (#30)
We have a lot of tests that rely on sleeping. That's never a good sign
in tests.

This change adds support for the client waiting for a particular
response to be received. That allows us to correctly wait for events
that happen within a single client reliably. There's still a few
lingering sleeps, those require a different synchronization primitive.