Add `unordered` (#32)
We have a few tests that are inherently racy: they don't have any
synchronization primitives inside and furthermore _cannot_ have any
synchronization inside them.

This change allows us to tag a few tests as being `unordered`. this
means that their outputs will be compared as sets instead of as arrays.